Switch task queue to procrastinate based on psql

+++ aleph/tests/util.py
@@ -14,6 +14,8 @@ from flask import json
 from followthemoney import model
 from followthemoney.cli.util import read_entity
 from ftmq.store.fragments import get_store
+from openaleph_procrastinate.manage.db import get_db
+from servicelayer import env
 from servicelayer import settings as sls
 from sqlalchemy import text
 from werkzeug.utils import cached_property
@@ -235,8 +237,11 @@ class TestCase(unittest.TestCase):
             SETTINGS._global_test_state = True
             destroy_db()
             flask_migrate.upgrade()
-            delete_index()
-            upgrade_search()
+            if not env.to_bool("TESTING_KEEP_INDEX"):
+                delete_index()
+                upgrade_search()
+            procrastinate_db = get_db()
+            procrastinate_db.configure()
         else:
             clear_index()
             for table in reversed(db.metadata.sorted_tables):
@@ -252,7 +257,7 @@ class TestCase(unittest.TestCase):
         db.session.rollback()
         db.session.close()
 
-        ftm_store = get_store(SETTINGS.DATABASE_URI)
+        ftm_store = get_store()
         for dataset in ftm_store.all():
             dataset.delete()
